debatebrawl_backend/requirements.txt
fastapi
uvicorn[standard]
firebase-admin
requests
pygad
pydantic<2.0.0
python-dotenv
numpy
pandas
matplotlib
seaborn
typing-extensions
python-multipart